% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleDirectedNetwork}
\alias{SampleDirectedNetwork}
\title{Linear network representing a directed road structure}
\format{An object of class \code{SpatialLinesDataFrame} with 2513 rows and 3 columns.}
\usage{
SampleDirectedNetwork
}
\description{
An \code{SpatialLinesDataFrame} representing a directed road structure. The \code{data.frame} attached has three columns, \code{V1}, \code{V2} and \code{Dir} that determine the direction of the network. The column \code{Dir} indicates the type of flow existing between \code{V1} and \code{V2}: none (0), from \code{V1} to \code{V2} (1), from \code{V2} to \code{V1} (-1) or bidirectional (2).
}
\keyword{datasets}
