% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleNetwork}
\alias{SampleNetwork}
\title{Linear network representing a road structure}
\format{An object of class \code{SpatialLines} of length 2513.}
\usage{
SampleNetwork
}
\description{
An \code{SpatialLines} representing a road structure
}
\keyword{datasets}
