% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SamplePointPattern}
\alias{SamplePointPattern}
\title{Point pattern on a road network simulating traffic accidents}
\format{An object of class \code{lpp} (inherits from \code{ppx}) of length 3.}
\usage{
SamplePointPattern
}
\description{
An \code{lpp} object generated with a Poisson process of intensity 0.01 over the \code{SampleNetwork} road structure. The pattern is marked with three categorical variables: \code{accident.date}, \code{collision} and \code{vehicles}, which were also randomly constructed
}
\keyword{datasets}
